Proxy para servicio de pagos
